Modified DeleteInstanceAction such that a namePattern can be passed

+++ src/main/java/org/k8loud/executor/actions/openstack/DeleteInstanceAction.java
@@ -7,6 +7,7 @@ import org.k8loud.executor.exception.ActionException;
 import org.k8loud.executor.exception.OpenstackException;
 import org.k8loud.executor.exception.ValidationException;
 import org.k8loud.executor.openstack.OpenstackService;
+import static org.k8loud.executor.util.Util.resultMap;
 
 import java.util.Collections;
 import java.util.List;
@@ -15,7 +16,7 @@ import java.util.Map;
 @EqualsAndHashCode
 public class DeleteInstanceAction extends OpenstackAction {
     private String region;
-    private String name;
+    private String namePattern;
     private List<String> serverIds;
 
     public DeleteInstanceAction(Params params, OpenstackService openstackService) throws ActionException {
@@ -23,26 +24,29 @@ public class DeleteInstanceAction extends OpenstackAction {
     }
 
     @Builder
-    public DeleteInstanceAction(OpenstackService openstackService, String region, String name, List<String> serverIds) {
+    public DeleteInstanceAction(OpenstackService openstackService, String region, String namePattern,
+                                List<String> serverIds) {
         super(openstackService);
         this.region = region;
-        this.name = name;
+        this.namePattern = namePattern;
         this.serverIds = serverIds;
     }
 
     @Override
     public void unpackParams(Params params) {
         region = params.getRequiredParam("region");
-        name = params.getRequiredParam("name");
+        namePattern = params.getOptionalParam("namePattern", null);
         serverIds = params.getOptionalParamAsListOfStrings("serverIds", Collections.emptyList());
     }
 
     @Override
     protected Map<String, String> executeBody() throws OpenstackException, ValidationException {
-        if (serverIds.isEmpty()){
-            return openstackService.deleteServers(region, name);
+        if (namePattern == null && serverIds.isEmpty()) {
+            return resultMap("Either namePattern or serverIds should be provided");
         }
-
+        if (serverIds.isEmpty()) {
+            return openstackService.deleteServers(region, namePattern);
+        }
         return openstackService.deleteServers(region, serverIds);
     }
 }
